+++ packages/big-design/src/components/Dropdown/index.ts
@@ -1,1 +1,2 @@
-export { Dropdown, DropdownItem, DropdownLinkItem, DropdownProps } from './Dropdown';
+export { Dropdown } from './Dropdown';
+export { DropdownItem, DropdownLinkItem, DropdownProps } from './types';
